Make recalc from csv work

+++ app/Helpers/OrdersRecalculatorBasedOnPeriod.php
@@ -77,11 +77,13 @@ class OrdersRecalculatorBasedOnPeriod
         $orderItemsValueWithTransport = $order->getItemsGrossValue() + $order->shipment_price_for_us;
         $totalPaymentsBuying = $order->payments->where('operation_type', 'Wpłata/wypłata bankowa - związana z fakturą zakupową')->sum('amount');
 
-        $arr = [];
-        if ($orderItemsValueWithTransport !== $totalPaymentsBuying) {
-            AddLabelService::addLabels($order, [258], $arr, []);
-        } else {
-            RemoveLabelService::removeLabels($order, [258],  $arr, [], []);
+        if ($order->payments->where('operation_type', 'Wpłata/wypłata bankowa - związana z fakturą zakupową')->first()) {
+            $arr = [];
+            if ($orderItemsValueWithTransport !== $totalPaymentsBuying) {
+                AddLabelService::addLabels($order, [258], $arr, []);
+            } else {
+                RemoveLabelService::removeLabels($order, [258],  $arr, [], []);
+            }
         }
     }
 }
